refactor(web) - stop rendering project pages conditonally
- migrate to actual files to represent the project pages
- replace anchor tag hrefs with nextjs Link components
