+++ web/hoc/ProjectLayout/ProjectLayout.tsx
@@ -12,7 +12,7 @@ export interface ProjectLayoutProps {
 export const ProjectLayout: React.FC<ProjectLayoutProps> = ({ project, selected, children }) => {
     return (
         <div className="flex w-screen h-screen justify-between items-center">
-            <Sidebar {...{selected, project}} />
+            <Sidebar selected={selected} />
             <div className="flex flex-col w-full h-screen justify-between items-center">
                 <Navbar avatar="/avatar.png" />
                 <div className="flex flex-col w-full h-full justify-center items-center pb-20 pr-64">{children}</div>
